requirements.txt
torch
torchvision
numpy
pandas
matplotlib
albumentations
opencv-python
Pillow